src/requirements.txt
Pillow==10.1.0
tk==0.1.0
